Change the initial robot position
Spawn model node is called with initial joint parameters
preset. However, there is a bug with the spawn model on
gazebo. The fix was to first pause the simulation at the
initialization, spawn the model with the updated coords.
and the unpause again to have the joint formations change.

This is tested on ur5e cartesian robot.